Revert "adding textArea & fiex input and buton components"
