--- src/assets/DropUpIcon.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export const UpIcon = () => {
-    return (
-        <svg xmlns="http://www.w3.org/2000/svg" width="12" height="12" viewBox="0 0 24 24">
-            <path d="M0 16.67l2.829 2.83 9.175-9.339 9.167 9.339 2.829-2.83-11.996-12.17z" />
-        </svg>
-    );
-};
