+++ src/components/Button/Button.css
@@ -1,21 +1,11 @@
-* {
-    font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-}
-
 .commonButton {
     padding: 8px 12px;
     border: none;
+    border-radius: 5px;
     font-size: 16px;
-    background-color: #0094DA;
+    background-color: #029cca;
     color: white;
     cursor: pointer;
-    width: 357px;
-    height: 41px;
-    border-radius: 6px;
-}
-
-.commonButton:hover {
-    background-color: #9E9E9E;
 }
 
 .commonButton:hover {
