+++ src/components/Input/Input.css
@@ -1,110 +1,14 @@
-.inputField {
-    position: relative;
+.commonInput {
+    width: 50%;
+    margin-bottom: 1rem;
+    margin-top: 1rem;
+    border-radius: 0.25rem;
+    border: 1px solid #ccc;
+    padding: 0.5rem;
+    font-size: 1rem;
 }
 
-input[type='text'],
-input[type='email'] {
-    padding: 8px 20px;
-    border: 1px solid rgba(147, 128, 108, 0.25);
-    border-radius: 6px;
-    width: 357px;
-    height: 30px;
-    color: #000000;
-    font-size: 16px;
-    line-height: 16px;
-    font-weight: 400;
-}
-
-input[type='password'] {
-    padding: 8px 20px;
-    border: 1px solid rgba(147, 128, 108, 0.25);
-    border-radius: 6px;
-    width: 357px;
-    height: 30px;
-    color: #000000;
-    font-size: 16px;
-    font-weight: 400;
-    line-height: 16px;
-}
-
-.inputField input:focus {
-    border: 1.8px solid #6b6b6b;
+.commonInput:focus {
     outline: none;
-}
-
-span.togglePasswordVisibility {
-    float: right;
-    margin-left: -50px;
-    margin-top: 12px;
-    position: absolute;
-}
-
-span.togglePlaceholderVisibility {
-    pointer-events: none;
-    height: 16.23px;
-    width: 245.69px;
-    opacity: 0.5;
-    color: #6b6b6b;
-    position: absolute;
-    padding-left: 20px;
-    left: 0px;
-    top: 15px;
-    font-style: normal;
-    font-weight: 400;
-    font-size: 18px;
-    line-height: 19px;
-    cursor: text;
-}
-
-.inputField input:not(:placeholder-shown)+span.togglePlaceholderVisibility,
-.inputField input:focus+span.togglePlaceholderVisibility {
-    top: 2px;
-    left: -15px;
-    padding: 0px;
-    color: #9E9E9E;
-    -webkit-transform: scale(0.7) translateY(-10%) translateX(-8.5px);
-    transform: scale(0.7) translateY(-10%) translateX(-8.5px);
-}
-
-/* Code for the inline span and input Field */
-.InputAddOn {
-    display: flex;
-    margin-bottom: 1.5em;
-}
-
-input.InputAddOn-field {
-    color: #000000;
-    width: 330px;
-    font-weight: 400;
-}
-
-.InputAddOn-field:not(:first-child) {
-    border-left: 0;
-}
-
-.InputAddOn-field:not(:last-child) {
-    border-right: 0;
-}
-
-.InputAddOn-item {
-    background-color: rgba(147, 128, 108, 0.1);
-    color: #666666;
-    font: inherit;
-    font-weight: normal;
-}
-
-.InputAddOn-field,
-.InputAddOn-item {
-    border: 1px solid rgba(147, 128, 108, 0.25);
-    padding: 14px 15px;
-}
-
-.InputAddOn-field:first-child,
-.InputAddOn-item:first-child {
-    border-radius: 6px 0 0 6px;
-}
-
-.InputAddOn-field:last-child,
-.InputAddOn-item:last-child {
-    border-radius: 0 6px 6px 0;
+    border: 2px solid #029cca;
 }
