--- src/components/TextArea/TextArea.css
@@ -1,13 +1,0 @@
-.textField {
-  background-color: rgba(147, 128, 108, 0.1);
-  color: black;
-  padding: 6px;
-  font-weight: normal;
-  border: 1px solid rgba(147, 128, 108, 0.25);
-  border-radius: 6px;
-}
-
-.textField:focus {
-  border: 1.8px solid #6b6b6b;
-  outline: none;
-}
