--- src/components/TextArea/index.ts
@@ -1,1 +1,0 @@
-export * from './TextArea';
