+++ src/index.ts
@@ -8,5 +8,3 @@ export * from './components/Alert';
 export * from './components/Tooltip';
 export * from './components/AuthHoc';
 export * from './components/Sidebar';
-export * from './components/NavBar';
-export * from './components/TextArea';
