+++ src/stories/Select.stories.tsx
@@ -20,15 +20,3 @@ Primary.args = {
         console.log(val);
     }
 };
-
-export const Single = Template.bind({});
-
-Single.args = {
-    options: ['Green', 'Red', 'Yellow', 'Blue', 'Pink', 'Black', 'Orange'],
-    placeholder: 'Select...',
-    // isMulti: false,
-    isSearchable: true,
-    onChange: function (val: any): void {
-        console.log(val);
-    }
-};
